+++ src/Domain/ValueObject/EmailAddress.php
@@ -1,0 +1,38 @@
+<?php
+
+/*
+ * This file is part of the eluceo/iCal package.
+ *
+ * (c) 2021 Markus Poerschke <markus@poerschke.nrw>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Eluceo\iCal\Domain\ValueObject;
+
+use InvalidArgumentException;
+
+final class EmailAddress
+{
+    private string $emailAddress;
+
+    public function __construct(string $emailAddress)
+    {
+        if (!filter_var($emailAddress, FILTER_VALIDATE_EMAIL)) {
+            throw new InvalidArgumentException("$emailAddress is no valid e-mail address");
+        }
+
+        $this->emailAddress = $emailAddress;
+    }
+
+    public function getEmailAddress(): string
+    {
+        return $this->emailAddress;
+    }
+
+    public function toUri(): Uri
+    {
+        return new Uri('mailto:' . urlencode($this->emailAddress));
+    }
+}